✨ feat(新增站内通知等功能、修复部分BUG):
1.简单添加站内通知 2.修改部分BUG 3.修改server 开发环境构建方式，不再使用自带的watch

+++ server/scripts/dev/remove.ts
@@ -1,6 +1,8 @@
 import fs from "fs";
 function remove(path: string) {
   let targetPath = `dist/${path.replace(".ts", ".js")}`;
-  fs.unlinkSync(targetPath);
+  if (fs.existsSync(targetPath)) {
+    fs.unlinkSync(targetPath);
+  }
 }
 export default remove;
